refactor: protocol index.ts to export.ts (also in subdirectories)
