+++ packages/protocol/package.json
@@ -7,8 +7,8 @@
     "url": "git+https://github.com/streamr-dev/network-monorepo.git",
     "directory": "packages/protocol"
   },
-  "main": "dist/src/index.js",
-  "types": "dist/src/index.d.ts",
+  "main": "dist/src/exports.js",
+  "types": "dist/src/exports.d.ts",
   "scripts": {
     "build": "tsc -b tsconfig.node.json",
     "prepublishOnly": "npm run clean && NODE_ENV=production npm run build",
