--- packages/protocol/src/index.ts
@@ -1,5 +1,0 @@
-export * from "./protocol/control_layer/index"
-export * from "./protocol/message_layer/index"
-export * from "./protocol/tracker_layer/index"
-export * from "./errors/index"
-export * from "./utils/index"
