+++ packages/protocol/src/protocol/control_layer/proxy_connection_response/ProxyConnectionResponse.ts
@@ -6,7 +6,7 @@ import {
     validateIsString
 } from '../../../utils/validations'
 import { StreamID } from '../../../utils/StreamID'
-import { StreamPartID, toStreamPartID } from "../../../utils"
+import { StreamPartID, toStreamPartID } from "../../../utils/exports"
 import { ProxyDirection } from '../../../utils/types'
 
 interface Options extends ControlMessageOptions {
