+++ packages/protocol/src/protocol/tracker_layer/instruction_message/InstructionMessage.ts
@@ -4,8 +4,8 @@ import {
     validateIsArray
 } from '../../../utils/validations'
 import TrackerMessage, { TrackerMessageOptions } from '../TrackerMessage'
-import { StreamID } from '../../../utils'
-import { StreamPartID } from "../../../utils"
+import { StreamID } from '../../../utils/exports'
+import { StreamPartID } from "../../../utils/exports"
 import { toStreamPartID } from "../../../utils/StreamPartID"
 
 export interface Options extends TrackerMessageOptions {
