+++ packages/protocol/src/protocol/tracker_layer/status_ack_message/StatusAckMessage.ts
@@ -1,7 +1,7 @@
 import { validateIsNotEmptyString, validateIsNotNegativeInteger } from '../../../utils/validations'
 import TrackerMessage, { TrackerMessageOptions } from '../TrackerMessage'
-import { StreamID } from '../../../utils'
-import { StreamPartID } from "../../../utils"
+import { StreamID } from '../../../utils/exports'
+import { StreamPartID } from "../../../utils/exports"
 import { toStreamPartID } from "../../../utils/StreamPartID"
 
 export interface Options extends TrackerMessageOptions {
