+++ packages/protocol/test/unit/exports.test.ts
@@ -1,11 +1,11 @@
-import * as Protocol from '../../src'
-import * as ControlLayer from "../../src/protocol/control_layer/index"
-import * as MessageLayer from "../../src/protocol/message_layer/index"
-import * as TrackerLayer from "../../src/protocol/tracker_layer/index"
-import * as Errors from "../../src/errors/index"
-import * as Utils from "../../src/utils/index"
+import * as Protocol from '../../src/exports'
+import * as ControlLayer from "../../src/protocol/control_layer/exports"
+import * as MessageLayer from "../../src/protocol/message_layer/exports"
+import * as TrackerLayer from "../../src/protocol/tracker_layer/exports"
+import * as Errors from "../../src/errors/exports"
+import * as Utils from "../../src/utils/exports"
 
-describe('index.ts', () => {
+describe('exports.ts', () => {
     it('exports all members of containers at top level also ensuring there are no duplicate names', () => {
         const containers = [ControlLayer, MessageLayer, TrackerLayer, Errors, Utils]
         const numKeys = containers.map((value) => Object.keys(value).length).reduce((a, b) => a + b)
