+++ packages/protocol/test/unit/protocol/control_layer/BroadcastMessageSerializerV2.test.ts
@@ -7,7 +7,7 @@ import {
     ContentType,
     EncryptionType,
     StreamMessageType
-} from '../../../../src/index'
+} from '../../../../src/exports'
 import { toEthereumAddress } from '@streamr/utils'
 import { SIGNATURE_TYPE_ETH } from '../../../../src/protocol/message_layer/StreamMessageSerializerV32'
