+++ packages/protocol/test/unit/protocol/control_layer/ProxyConnectionRequest.test.ts
@@ -1,7 +1,7 @@
 import assert from 'assert'
 
 import ValidationError from '../../../../src/errors/ValidationError'
-import { ProxyConnectionRequest, ControlMessage, toStreamID } from '../../../../src/index'
+import { ProxyConnectionRequest, ControlMessage, toStreamID } from '../../../../src/exports'
 import { ProxyDirection } from '../../../../src/utils/types'
 
 describe('ProxyConnectionRequest', () => {
