+++ packages/protocol/test/unit/protocol/control_layer/UnsubscribeRequest.test.ts
@@ -3,7 +3,7 @@ import assert from 'assert'
 import UnsubscribeRequest from '../../../../src/protocol/control_layer/unsubscribe_request/UnsubscribeRequest'
 import ControlMessage from '../../../../src/protocol/control_layer/ControlMessage'
 import ValidationError from '../../../../src/errors/ValidationError'
-import { toStreamID } from '../../../../src'
+import { toStreamID } from '../../../../src/exports'
 
 describe('UnsubscribeRequest', () => {
     describe('constructor', () => {
