+++ packages/protocol/test/unit/protocol/tracker_layer/StatusAckMessage.test.ts
@@ -2,7 +2,7 @@ import assert from 'assert'
 
 import ValidationError from '../../../../src/errors/ValidationError'
 import TrackerMessage from '../../../../src/protocol/tracker_layer/TrackerMessage'
-import { toStreamID } from '../../../../src'
+import { toStreamID } from '../../../../src/exports'
 import StatusAckMessage from '../../../../src/protocol/tracker_layer/status_ack_message/StatusAckMessage'
 
 describe('StatusAckMessage', () => {
